moves reusable functions to services
